linreg: Split core C to separate file
